Create vector.py
The Vector class contains: finding the length of a vector, the scalar product of vectors, checking for collinearity of vectors, finding the cosine between vectors. Arithmetic operators have also been reloaded, which allows operations between two objects of the Vector class (for multiplication and division, it is possible to find the product/division of a vector by a scalar).